update imports to remove circularity
