+++ greenheart/simulation/technologies/hydrogen/electrolysis/PEM_BOP/PEM_BOP.py
@@ -2,7 +2,7 @@ import os
 import numpy as np
 import pandas as pd
 import scipy.optimize
-import greenheart.tools.eco.electrolysis as elec # import for get_electrolyzer_BOL_efficiency function
+from greenheart.simulation.technologies.hydrogen.electrolysis.PEM_tools import get_electrolyzer_BOL_efficiency
 
 file_path = os.path.dirname(os.path.abspath(__file__))
 
@@ -95,7 +95,6 @@ def pem_bop(
     Returns:
         energy_consumption_bop_kwh (list or np.array): Energy consumed by electrolyzer BOP in kWh.
     """
-    
     operating_ratios = power_profile_to_electrolyzer_kw / (electrolyzer_rated_mw * 1e3)
 
     curve_coeff, min_ratio, max_ratio, min_efficiency, max_efficiency = (
@@ -116,7 +115,7 @@ def pem_bop(
         max_efficiency,
     )
 
-    BOL_efficiency = elec.get_electrolyzer_BOL_efficiency()  # kwh/kg
+    BOL_efficiency = get_electrolyzer_BOL_efficiency()  # kwh/kg
 
     BOL_kg = (electrolyzer_rated_mw * 1000) / BOL_efficiency  # kg/hr
